necessary updates

+++ Backend/RRSS/src/main/java/com/demo/rrss/rrssbackend/service/ProductService.java
@@ -8,22 +8,31 @@ import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 import org.springframework.ui.Model;
+
+import com.demo.rrss.rrssbackend.controller.UsersController;
 import com.demo.rrss.rrssbackend.entity.Product;
+import com.demo.rrss.rrssbackend.entity.Users;
 import com.demo.rrss.rrssbackend.repository.ProductRepository;
+import com.demo.rrss.rrssbackend.repository.UsersRepository;
 import com.demo.rrss.rrssbackend.rest.request.ProductRequest;
 
 @Service
 public class ProductService {
 	@Autowired
 	ProductRepository repository;
+	@Autowired
+	UsersRepository uRepository;
 
 	public Product getProduct(Long productId) {
 		return repository.findById(productId)
 				.orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Product not found"));
 	}
 
-	public void addProduct(ProductRequest request, Model model) { // TODO yetki kontrolü eklenecek
+	public void addProduct(ProductRequest request, Model model) {
 		Long userId = (Long) model.getAttribute("userId"); 
+		Users user = uRepository.findById(userId).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "User not found"));
+		if (!user.getIsMerchant())
+			throw new ResponseStatusException(HttpStatus.FORBIDDEN, "User is not a merchant");
 		Product product = new Product();
 		product.setCategoryId(request.getCategoryId());
 		product.setDescription(request.getDescription());
@@ -34,9 +43,10 @@ public class ProductService {
 		repository.save(product);
 	}
 
-	public void updateProduct(Long productId, ProductRequest request) { // TODO yetki kontrolü eklenecek
+	public void updateProduct(Long productId, ProductRequest request, Model model) {
+		Long userId = (Long) model.getAttribute("userId");
 		Optional<Product> existingProduct = repository.findById(productId);
-		if (existingProduct.isPresent()) {
+		if (existingProduct.isPresent() && existingProduct.get().getUserId() == userId){
 			Product product = existingProduct.get();
 			product.setCategoryId(request.getCategoryId());
 			product.setDescription(request.getDescription());
@@ -49,15 +59,16 @@ public class ProductService {
 		}
 	}
 
-	public void deleteProduct(Long productId) { // TODO yetki kontrolü eklenecek
-		if (repository.existsById(productId))
+	public void deleteProduct(Long productId, Model model) {
+		Long userId = (Long) model.getAttribute("userId");
+		if (repository.existsById(productId) && repository.findById(productId).get().getUserId() == userId)
 			repository.deleteById(productId);
 		else
 			throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Product not found");
 
 	}
 
-	public List<Product> getAllProducts(Long categoryId) { // TODO yetki kontrolü eklenecek
+	public List<Product> getAllProducts(Long categoryId) {
 		if (categoryId == -1)
 			return repository.findAllMax50();
 		else {
